testing and/or fixing minimap. getTileHeight/Width was probably always returning 0.

+++ src/main/java/fauxpas/views/MiniMapWorldView.java
@@ -55,11 +55,11 @@ public class MiniMapWorldView {
     }
 
     private int getTileWidth() {
-        return (this.width/(world.getWidth()*assets.getTileDimension()))*assets.getTileDimension();
+        return (int)((float)this.width/(world.getWidth()*assets.getTileDimension()))*assets.getTileDimension();
     }
 
     private int getTileHeight() {
-        return (this.height/(world.getHeight()*assets.getTileDimension()))*assets.getTileDimension();
+        return (int)((float)this.height/(world.getHeight()*assets.getTileDimension()))*assets.getTileDimension();
     }
 
 }
